Update Metadata file structure
