+++ context/app/static/js/components/Detail/Metadata/index.js
@@ -1,0 +1,3 @@
+import Metadata from './Metadata';
+
+export default Metadata;
