+++ context/app/static/js/components/Detail/Metadata/style.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+import Typography from '@material-ui/core/Typography';
+import Paper from '@material-ui/core/Paper';
+
+const StyledTypography = styled(Typography)`
+  margin: 2px 0px 2px 0px;
+`;
+
+const FlexPaper = styled(Paper)`
+  display: flex;
+  padding: 30px 40px 30px 40px;
+`;
+
+export { StyledTypography, FlexPaper };
